[ADD] Option -m. -C is kept with a pipe.

+++ includes/ls.h
@@ -31,16 +31,16 @@ typedef struct			s_file
 
 
 int						ft_ls(int ac, char **av);
-int						parse_ls_options(int ac, char **av, int *opt, int *real_args);
-int						parse_option_line(char *av, int *opt);
-t_dlist					*analyze_args(int ac, char **av, int *opt);
-int						analyze_file(char *file, int opt);
+int						parse_ls_options(int ac, char **av, unsigned long long *opt, int *real_args);
+int						parse_option_line(char *av, unsigned long long *opt);
+t_dlist					*analyze_args(int ac, char **av, unsigned long long *opt);
+int						analyze_file(char *file, unsigned long long opt);
 void					print_file(t_stat file_stats, char *file,
-t_ls_padding padding, int opt);
+t_ls_padding padding, unsigned long long opt);
 void					print_file_name(t_stat file_stats, char *file,
-size_t padding, int opt);
-void					get_ls_time(char *res, t_stat stats, int opt);
-void					print_size(off_t size, int padding, int opt);
+size_t padding, unsigned long long opt);
+void					get_ls_time(char *res, t_stat stats, unsigned long long opt);
+void					print_size(off_t size, int padding, unsigned long long opt);
 void					print_size_short(off_t size);
 int						sort_files(t_list **lst, int start, int end);
 int						compare_none(void *s1, void *s2);
@@ -49,15 +49,15 @@ int						compare_names_no_case(void *s1, void *s2);
 int						compare_times(void *s1, void *s2);
 int						compare_atimes(void *s1, void *s2);
 int						compare_ctimes(void *s1, void *s2);
-void					print_dlist(t_dlist *dlst, int opt);
-void					print_dlist_reverse(t_dlist *dlst, int opt);
+void					print_dlist(t_dlist *dlst, unsigned long long opt);
+void					print_dlist_reverse(t_dlist *dlst, unsigned long long opt);
 void					free_t_file(void *file, size_t size);
-int						analyze_directory(char *file_name, int *opt);
-int						(*get_compare_func(int opt))(void*, void*);
+int						analyze_directory(char *file_name, unsigned long long *opt);
+int						(*get_compare_func(unsigned long long opt))(void*, void*);
 int						print_dlist_col(t_dlist *lst, size_t len,
-unsigned short winsize, int opt);
+unsigned short winsize, unsigned long long opt);
 int						is_arg_an_option_line(char *av);
-t_ls_padding			get_padding(t_dlist *dlst, blksize_t *dir_size, int opt);
+t_ls_padding			get_padding(t_dlist *dlst, blksize_t *dir_size, unsigned long long opt);
 int						get_doublelen(double size);
 
 #endif
