+++ src/print_size.c
@@ -183,7 +183,7 @@ void	print_size_readable_si(off_t long_size, int padding, double divider)
 **	Print file size
 */
 
-void	print_size(off_t size, int padding, int opt)
+void	print_size(off_t size, int padding, unsigned long long opt)
 {
 	if (opt & OPT_L || opt & OPT_G)
 	{
